style changes and made a component for counter
