+++ src/index.css
@@ -10,7 +10,9 @@
 * {
     font-family:'Lucida Grande', 'Lucida Sans', Arial, sans-serif;
     margin: 0;
+    padding: 0;
     scroll-margin-top: 80px;
+    outline: 1px solid red;
 }
 *::selection {
     background-color: #383F51;
@@ -626,7 +628,6 @@ footer form button:hover {
         position: fixed;
         width: 100%;
         top: 50;
-        padding-top: 5%;
         background-color: #4986B6;
     }
     .drop-down li {
@@ -751,7 +752,7 @@ footer form button:hover {
         display: block;
         background-color: #4986B6;
         position: fixed;
-        margin-top: 15%;
+        
     }
     .drop-down li {
         width: fit-content;
@@ -765,6 +766,9 @@ footer form button:hover {
     /*End of navbar section */
 
     /* Landing Section */
+    .typewriter-text {
+        display: none;
+    }
     .pfp {
         width: 60%;
         display: block;
